Move Player & Token into separate comps, imported into Board through getChildren.tsx
