+++ src/components/Board/Board.module.scss
@@ -43,120 +43,3 @@
 .path {
   @include bg-image-layout();
 }
-
-.player {
-  width: 20px;
-  height: 20px;
-  border-radius: 100%;
-  position: absolute;
-  z-index: 1;
-
-  &.e1 {
-    background-color: $enemy1-main-color;
-    border: 1px solid $enemy1-secondary-color;
-  }
-
-  &.e2 {
-    background-color: $enemy2-main-color;
-    border: 1px solid $enemy2-secondary-color;
-  }
-
-  &.0 {
-    background-color: $player0-main-color;
-    border: 1px solid $player0-secondary-color;
-  }
-
-  &.1 {
-    background-color: $player1-main-color;
-    border: 1px solid $player1-secondary-color;
-  }
-
-  &.2 {
-    background-color: $player2-main-color;
-    border: 1px solid $player2-secondary-color;
-  }
-
-  &.3 {
-    background-color: $player3-main-color;
-    border: 1px solid $player3-secondary-color;
-  }
-
-  &.4 {
-    background-color: $player4-main-color;
-    border: 1px solid $player4-secondary-color;
-  }
-
-  &.5 {
-    background-color: $player5-main-color;
-    border: 1px solid $player5-secondary-color;
-  }
-}
-
-.enemyPlayerWrapper {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .triangle {
-    position: absolute;
-    z-index: 1;
-
-    width: 0;
-    height: 0;
-
-    $triangle-width: 4px;
-    $triangle-height: $triangle-width * 2 * 0.866;
-
-    &.up {
-      margin-top: -30px;
-      border-left: $triangle-width solid transparent;
-      border-right: $triangle-width solid transparent;
-
-      &.e1 {
-        border-bottom: $triangle-height solid $enemy1-secondary-color;
-      }
-      &.e2 {
-        border-bottom: $triangle-height solid $enemy2-main-color;
-      }
-    }
-
-    &.down {
-      margin-bottom: -30px;
-      border-left: $triangle-width solid transparent;
-      border-right: $triangle-width solid transparent;
-
-      &.e1 {
-        border-top: $triangle-height solid $enemy1-secondary-color;
-      }
-      &.e2 {
-        border-top: $triangle-height solid $enemy2-main-color;
-      }
-    }
-
-    &.right {
-      margin-right: -32px;
-      border-top: $triangle-width solid transparent;
-      border-bottom: $triangle-width solid transparent;
-
-      &.e1 {
-        border-left: $triangle-height solid $enemy1-secondary-color;
-      }
-      &.e2 {
-        border-left: $triangle-height solid $enemy2-main-color;
-      }
-    }
-
-    &.left {
-      margin-left: -30px;
-      border-top: $triangle-width solid transparent;
-      border-bottom: $triangle-width solid transparent;
-
-      &.e1 {
-        border-right: $triangle-height solid $enemy1-secondary-color;
-      }
-      &.e2 {
-        border-right: $triangle-height solid $enemy2-main-color;
-      }
-    }
-  }
-}
