+++ src/components/Board/PositionChildren/Token/Token.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Token = ({ type }: { type: 'sight' | 'sound' }) => {
+  return <div className={`${type}-token`} />;
+};
+
+export default Token;
